apps/rag-ba-ocr/pyproject.toml
[project]
name = "rag-ba-ocr"
version = "0.1.0"
description = ""
authors = [
    {name = "MarsBroshok96",email = "marsbroshok1996@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "paddleocr (>=3.4.0,<4.0.0)",
    "opencv-python-headless (>=4.13.0.92,<5.0.0.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "shapely (>=2.1.2,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "protobuf (>=3.20.2,<5)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false